data-analysis-service/requirements.txt
fastapi>=0.68.1
uvicorn>=0.15.0
sqlalchemy>=1.4.23
asyncpg>=0.24.0
psycopg2-binary>=2.9.1
python-dotenv>=0.19.0
pydantic>=1.8.2
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
kafka-python>=2.0.2
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.5
requests==2.26.0
aiohttp==3.8.1
plotly==5.18.0
dash==2.14.2
dash-bootstrap-components==1.5.0